refactor(frontend/src/components/pages
/Filter/FilterCmd.jsx):

extract handleCopyClick function into

Command.jsx component
